Move and test DriverSlowTrigger
