+++ src/main/java/com/systemmeltdown/robotlib/triggers/DriverSlowTrigger.java
@@ -1,21 +1,26 @@
-package common;
+package com.systemmeltdown.robotlib.triggers;
 
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj.GenericHID.Hand;
 import edu.wpi.first.wpilibj.buttons.Trigger;
-import common.RobotMap;
 
 public class DriverSlowTrigger extends Trigger {
   private XboxController controller;
+  private double triggerThreshold;
 
   public DriverSlowTrigger(XboxController controller) {
+    this(controller, 0.25);
+  }
+
+  public DriverSlowTrigger(XboxController controller, double triggerThreshold) {
     this.controller = controller;
+    this.triggerThreshold = triggerThreshold;
   }
 
   @Override
   public boolean get() {
     double axis = controller.getTriggerAxis(Hand.kRight);
 
-    return (axis > RobotMap.DRIVER_SLOW_TRIGGER_THRESHOLD);
+    return (axis > triggerThreshold);
   }
 }
